[TAG LABELS] refactoring of tag label
